requirements.txt
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
chromadb==0.5.3
openai==1.55.0
httpx==0.27.2
PyMuPDF==1.23.8
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.30.0
python-dotenv==1.0.0
boto3==1.34.0
python-multipart==0.0.6
pydantic==2.5.3
requests==2.31.0